+++ CHANGELOG.md
@@ -18,6 +18,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Bug Fixes
 
+* Only check for valid Swift version for pod targets that use Swift  
+  [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
+  [#6733](https://github.com/CocoaPods/CocoaPods/pull/6733) 
+
 * Fix pod install error from 1.2.1 when working with static lib-only projects.  
   [Ben Asher](https://github.com/benasher44)
   [#6673](https://github.com/CocoaPods/CocoaPods/issues/6673)